+++ cloud/blockstore/libs/storage/disk_agent/actors/multi_agent_write_handler.h
@@ -8,6 +8,22 @@ namespace NCloud::NBlockStore::NStorage {
 
 ////////////////////////////////////////////////////////////////////////////////
 
+// IMultiAgentWriteHandler interface defines a method to perform write blocks
+// using multiple agents, returning a future with the response.
+class IMultiAgentWriteHandler
+{
+public:
+    virtual ~IMultiAgentWriteHandler() = default;
+
+    virtual NThreading::TFuture<
+        TEvDiskAgentPrivate::TMultiAgentWriteDeviceBlocksResponse>
+    PerformMultiAgentWrite(
+        TCallContextPtr callContext,
+        std::shared_ptr<NProto::TWriteDeviceBlocksRequest> request) = 0;
+};
+
+using IMultiAgentWriteHandlerPtr = std::shared_ptr<IMultiAgentWriteHandler>;
+
 IMultiAgentWriteHandlerPtr CreateMultiAgentWriteHandler(
     NActors::TActorSystem* actorSystem,
     NActors::TActorId diskAgentId);
