+++ cloud/blockstore/libs/storage/disk_agent/disk_agent_private.h
@@ -267,20 +267,4 @@ struct TEvDiskAgentPrivate
     BLOCKSTORE_DECLARE_EVENTS(UpdateSessionCache)
 };
 
-// IMultiAgentWriteHandler interface defines a method to perform write blocks
-// using multiple agents, returning a future with the response.
-class IMultiAgentWriteHandler
-{
-public:
-    virtual ~IMultiAgentWriteHandler() = default;
-
-    virtual NThreading::TFuture<
-        TEvDiskAgentPrivate::TMultiAgentWriteDeviceBlocksResponse>
-    PerformMultiAgentWrite(
-        TCallContextPtr callContext,
-        std::shared_ptr<NProto::TWriteDeviceBlocksRequest> request) = 0;
-};
-
-using IMultiAgentWriteHandlerPtr = std::shared_ptr<IMultiAgentWriteHandler>;
-
 }   // namespace NCloud::NBlockStore::NStorage
